requirements.txt
asgiref==3.11.0
av==16.1.0
beautifulsoup4==4.14.3
click==8.3.1
cloup==3.0.8
colorama==0.4.6
decorator==5.2.1
Django==6.0
djangorestframework==3.16.1
glcontext==3.0.0
isosurfaces==0.1.2
manim==0.19.2
ManimPango==0.6.1
mapbox_earcut==2.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
networkx==3.6.1
numpy==2.4.0
pillow==12.0.0
pycairo==1.29.0
pydub==0.25.1
pyglet==2.1.11
pyglm==2.8.3
Pygments==2.19.2
rich==14.2.0
scipy==1.16.3
screeninfo==0.8.1
skia-pathops==0.9.1
soupsieve==2.8.1
sqlparse==0.5.5
srt==3.5.3
svgelements==1.9.6
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
watchdog==6.0.0
